[HOTFIX-11-2-Check-In-MS-표현계층-형식변경] HelpSelect 관련 변경 테스트파일
